refact: move react to visitors
